Add a do_yield alias for the yield function
In strict mode, `yield` is a reserved keyword so a function in a module
(which is in strict mode) can not use yield as its identifier.

This gets around this issue by defining the function with the identifier
`do_yield` instead while exporting `yield` as an alias of `do_yield` so
that yield will still be an available for modules that expect to import
it.